Postgresql

+++ src/main/kotlin/com/example/data/models/UserReviews.kt
@@ -3,4 +3,4 @@ package com.example.data.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserReviews(val reviews: MutableList<Float>)
+data class UserReviews(val reviews: MutableList<Double>)
