+++ src/main/kotlin/com/example/data/requests/NewReviewRequest.kt
@@ -3,4 +3,4 @@ package com.example.data.requests
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class NewReviewRequest(val review: Float)
+data class NewReviewRequest(val review: Double)
