+++ src/main/kotlin/com/example/data/usersservice/UsersService.kt
@@ -3,9 +3,9 @@ package com.example.data.usersservice
 import com.example.data.models.UserReviews
 
 interface UsersService {
-    suspend fun insertNewReview(userId: String, review: Float): Boolean
+    suspend fun insertNewReview(userId: String, review: Double): Boolean
 
-    suspend fun calculateReview(userReviews: UserReviews?): Float
+    suspend fun calculateReview(userReviews: UserReviews?): Double
 
     suspend fun insertCalculatedReviews(): Boolean
